+++ src/main/java/proj_vendas/vendas/repository/Dados.java
@@ -8,20 +8,20 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import proj_vendas.vendas.model.Dado;
+import proj_vendas.vendas.model.empresa.Dado;
 
 @Transactional(readOnly = true)
 @Repository
 public interface Dados extends JpaRepository<Dado, Long>{
 
-	public Dado findByCodEmpresaAndData(int codEmpresa, String data);
+	public Dado findByCodEmpresaAndData(Long codEmpresa, String data);
 
-	public List<Dado> findByCodEmpresa(int codEmpresa);
+	public List<Dado> findByCodEmpresa(Long codEmpresa);
 
-	public List<Dado> findByCodEmpresaAndTrocoFinal(int codEmpresa, float trocoFinal);
+	public List<Dado> findByCodEmpresaAndTrocoFinal(Long codEmpresa, float trocoFinal);
 
-	public List<Dado> findByCodEmpresaAndDataBetween(int codEmpresa, String dataInicio, String dataFinal);
+	public List<Dado> findByCodEmpresaAndDataBetween(Long codEmpresa, String dataInicio, String dataFinal);
 
 	@Query("SELECT data FROM Dado u WHERE u.codEmpresa=:cod AND u.trocoFinal = 0")
-	public List<String> findByDiasEmAberto(@Param("cod") int codEmpresa);
+	public List<String> findByDiasEmAberto(@Param("cod") Long codEmpresa);
 }
