+++ src/main/java/proj_vendas/vendas/repository/Enderecos.java
@@ -8,14 +8,14 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import proj_vendas.vendas.model.Endereco;
+import proj_vendas.vendas.model.cadastros.Endereco;
 
 @Transactional(readOnly = true)
 @Repository
 public interface Enderecos extends JpaRepository<Endereco, Long>{
 
-	public List<Endereco> findByCodEmpresa(int codEmpresa);
+	public List<Endereco> findByCodEmpresa(Long codEmpresa);
 	
 	@Query("SELECT bairro, taxa FROM Endereco u WHERE u.codEmpresa=:cod")
-	public List<String> buscarEnderecos(@Param("cod") int codEmpresa);
+	public List<String> buscarEnderecos(@Param("cod") Long codEmpresa);
 }
