+++ src/main/java/proj_vendas/vendas/repository/Impressoes.java
@@ -6,14 +6,14 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import proj_vendas.vendas.model.ImpressaoMatricial;
+import proj_vendas.vendas.model.impressora.ImpressaoMatricial;
 
 @Transactional(readOnly = true)
 @Repository
 public interface Impressoes extends JpaRepository<ImpressaoMatricial, Long>{
 
-	public List<ImpressaoMatricial> findByCodEmpresaAndSetor(int codEmpresa, String setor);
+	public List<ImpressaoMatricial> findByCodEmpresaAndSetor(Long codEmpresa, String setor);
 
-	public List<ImpressaoMatricial> findByCodEmpresa(int codEmpresa);
+	public List<ImpressaoMatricial> findByCodEmpresa(Long codEmpresa);
 
 }
