+++ src/main/java/proj_vendas/vendas/repository/PedidoTemps.java
@@ -8,20 +8,20 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import proj_vendas.vendas.model.PedidoTemp;
+import proj_vendas.vendas.model.empresa.PedidoTemp;
 
 @Transactional(readOnly = true)
 @Repository
 public interface PedidoTemps extends JpaRepository<PedidoTemp, Long>{
 
-	public List<PedidoTemp> findByCodEmpresa(int codEmpresa);
+	public List<PedidoTemp> findByCodEmpresa(Long codEmpresa);
 
-	public List<PedidoTemp> findByCodEmpresaAndEnvio(int codEmpresa, String string);
+	public List<PedidoTemp> findByCodEmpresaAndEnvio(Long codEmpresa, String string);
 
-	public List<PedidoTemp> findByCodEmpresaAndDataAndComanda(int codEmpresa, String dia, Long comanda);
+	public List<PedidoTemp> findByCodEmpresaAndDataAndComanda(Long codEmpresa, String dia, Long comanda);
 	
 	@Query("SELECT COUNT(u) FROM PedidoTemp u WHERE u.codEmpresa=:cod AND u.data=:dia AND u.status=:status")
-    public int totalPedidos(@Param("cod") int codEmpresa, @Param("dia") String dia, @Param("status") String status);
+    public int totalPedidos(@Param("cod") Long codEmpresa, @Param("dia") String dia, @Param("status") String status);
 
-	public List<PedidoTemp> findByCodEmpresaAndStatus(int codEmpresa, String status);
+	public List<PedidoTemp> findByCodEmpresaAndStatus(Long codEmpresa, String status);
 }
